requirements.txt
paho-mqtt>=1.6.0
requests>=2.28.0
python-dotenv>=0.20.0
schedule>=1.1.0
pydantic>=2.0.0
